requirements.txt
absl-py==2.2.2
astunparse==1.6.3
attrs==25.3.0
Brotli @ file:///D:/bld/brotli-split_1725267609074/work
certifi @ file:///C:/b/abs_3beajm7umk/croot/certifi_1745939228545/work/certifi
cffi @ file:///D:/bld/cffi_1725560649097/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1746214863626/work
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.2
cvzone==1.6.1
cycler==0.12.1
filelock==3.19.1
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
graphviz==0.21
grpcio==1.71.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
h5py==3.13.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imageio==2.37.0
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.6
joblib==1.5.0
keras==3.9.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.9
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==4.25.7
psutil==7.0.0
py-cpuinfo==9.0.0
PyAudio==0.2.11
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydub @ file:///home/conda/feedstock_root/build_artifacts/pydub_1734638032476/work
pyFirmata==1.1.0
pyFirmata2==2.5.1
Pygments==2.19.1
pyparsing==3.2.3
pypiwin32==223
pyserial==3.5
PySocks @ file:///D:/bld/pysocks_1733217287171/work
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2025.2
pywin32==310
PyYAML==6.0.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rich==14.0.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
SpeechRecognition @ file:///home/conda/feedstock_root/build_artifacts/speechrecognition_1742707644995/work
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
tzdata==2025.2
ultralytics==8.3.184
ultralytics-thop==2.0.16
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1744323578849/work
Werkzeug==3.1.3
win_inet_pton @ file:///D:/bld/win_inet_pton_1733130564612/work
wrapt==1.17.2
zstandard==0.23.0
